pyproject.toml
[project]
name = "lattice-mcp"
version = "0.1.0"
description = "General-purpose lattice MCP server for persistent knowledge in AI coding assistants"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.83.14",
    "mcp>=1.27.1",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "rank-bm25>=0.2.2",
]

[project.scripts]
lattice-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lattice", "server.py"]

[dependency-groups]
dev = ["pytest>=8"]
